profiles_api2/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
get==2019.4.13
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
python3-openid==3.1.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
